explicit cast to avoid warning
git-svn-id: https://svn.r-project.org/R/trunk@71515 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/eval.c
@@ -5598,7 +5598,7 @@ static SEXP bcEval(SEXP body, SEXP rho, Rboolean useCache)
 	    R_bcstack_t *oldtop = R_BCNodeStackTop;
 	    RCNTXT *cntxt = BCNALLOC_CNTXT();
 	    BCNPUSH_INTEGER(GETOP());       /* pc offset for 'break' */
-	    BCNPUSH_INTEGER(pc - codebase); /* pc offset for 'next' */
+	    BCNPUSH_INTEGER((int)(pc - codebase)); /* pc offset for 'next' */
 	    if (is_for_loop) {
 		/* duplicate the for loop state data on the top of the stack */
 		R_bcstack_t *loopdata = oldtop - FOR_LOOP_STATE_SIZE;
